requirements.txt
aiogram==3.10.0
APScheduler==3.10.4
python-dotenv==1.0.1
pytz==2024.1

feedparser>=6.0.11
httpx>=0.27.0
trafilatura>=1.7.0
beautifulsoup4>=4.12.3
sumy>=0.11.0
simhash>=2.1.2